Extract code out of NeoForgeMod
To allow modders to easily extend potion fluid logic for their own potion fluids
