+++ src/main/java/net/neoforged/neoforge/fluids/potion/PotionFluid.java
@@ -1,0 +1,83 @@
+/*
+ * Copyright (c) NeoForged and contributors
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+
+package net.neoforged.neoforge.fluids.potion;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Fluid;
+import net.minecraft.world.level.material.FluidState;
+import net.minecraft.world.phys.Vec3;
+import net.minecraft.world.phys.shapes.Shapes;
+import net.minecraft.world.phys.shapes.VoxelShape;
+import net.neoforged.neoforge.common.NeoForgeMod;
+import net.neoforged.neoforge.fluids.FluidType;
+
+public class PotionFluid extends Fluid {
+    @Override
+    public Item getBucket() {
+        return NeoForgeMod.POTION_BUCKET.asItem();
+    }
+
+    @Override
+    protected boolean canBeReplacedWith(FluidState fluidState, BlockGetter level, BlockPos pos, Fluid fluid, Direction side) {
+        return true;
+    }
+
+    @Override
+    protected Vec3 getFlow(BlockGetter level, BlockPos pos, FluidState fluidState) {
+        return Vec3.ZERO;
+    }
+
+    @Override
+    public int getTickDelay(LevelReader level) {
+        return 0;
+    }
+
+    @Override
+    protected float getExplosionResistance() {
+        return 0F;
+    }
+
+    @Override
+    public float getHeight(FluidState fluidState, BlockGetter level, BlockPos pos) {
+        return 0F;
+    }
+
+    @Override
+    public float getOwnHeight(FluidState fluidState) {
+        return 0F;
+    }
+
+    @Override
+    protected BlockState createLegacyBlock(FluidState fluidState) {
+        return Blocks.AIR.defaultBlockState();
+    }
+
+    @Override
+    public boolean isSource(FluidState fluidState) {
+        return true;
+    }
+
+    @Override
+    public int getAmount(FluidState fluidState) {
+        return 0;
+    }
+
+    @Override
+    public VoxelShape getShape(FluidState fluidState, BlockGetter level, BlockPos pos) {
+        return Shapes.empty();
+    }
+
+    @Override
+    public FluidType getFluidType() {
+        return NeoForgeMod.POTION_TYPE.value();
+    }
+}
